restructuring condition
