+++ .homeycompose/app.json
@@ -1,6 +1,6 @@
 {
   "id": "app.picnic",
-  "version": "3.1.0",
+  "version": "3.2.0",
   "compatibility": ">=5.0.0",
   "sdk": 3,
   "name": {
