+++ app.json
@@ -1,7 +1,7 @@
 {
   "_comment": "This file is generated. Please edit .homeycompose/app.json instead.",
   "id": "app.picnic",
-  "version": "3.1.0",
+  "version": "3.2.0",
   "compatibility": ">=5.0.0",
   "sdk": 3,
   "name": {
@@ -176,10 +176,12 @@
       {
         "id": "groceries_ordered",
         "title": {
-          "en": "Groceries !{{have been|have not been}} ordered"
+          "en": "Groceries !{{have been|have not been}} ordered",
+          "nl": "Boodschappen !{{zijn besteld|zijn niet besteld}}"
         },
         "hint": {
-          "en": "Checks if an order with Picnic has been placed."
+          "en": "Check if there is an active order.",
+          "nl": "Kijk of er een actieve bestelling is."
         }
       }
     ],
