+++ lib/actions.js
@@ -8,7 +8,6 @@ class actions extends BaseClass {
   constructor({homey}) {
     super();
     this.homey = homey;
-    //this.utils = new utils({homey: this.homey});
   }
 
   async onInit() {
